extract antora config refresh
